Rework backend structure
User code now looks like this:

    from pyop2 import op2
    ...
    op2.init(backend="backend")
    # op2 calls

It is illegal to call an op2 function before op2.init.  To enforce
this, the selected backend is initially a void backend that raises a
NotImplementedError and reminds the user to call op2.init.

The user-facing API now just stubs user-visible calls and delegates to
the backend to do all the object creation.  A typical class now looks
like:

    class Kernel(object):
        def __new__(klass, *args):
            return backend.Kernel(*args)

On the backend side, in addition to the void backend, there is a
common backend that implements argument checking and defines
backend-agnostic stuff (such as access descriptors).  To implement a
new device-specific backend, classes should derive from this common
backend like so:

   from common import Access, IterationSpace, Set
   from common import READ, WRITE, RW, INC, MIN, MAX
   import common

   class Kernel(common.Kernel):
       def __init__(self, code, name):
           common.Kernel.__init__(self, code, name)
           # extra set-up here
       def compile(self):
           # compile kernel for device execution

An incomplete example of this is implemented in the cuda backend.