A sketch of a Cython interface to OP2-Common
The idea is that we only ever need to pass handles to the C data
structures around.  So we wrap the creation routines and keep hold of
handles to the C pointers.  When we need to call functions from the C
runtime library, the python objects are carrying around a handle to
the C pointer and can pass things appropriately.