Build C level op_arg types automatically
__call__ methods on datacarriers now return an Arg type, so that
multiple calls to access the same Dat (for instance) do not modify
it.  Necessary for building the set of unique arguments to a
par_loop.  The Arg type also knows how to build an op_arg C-level
struct describing the argument.  This is not called at initialisation
time, instead, if you want the handle, you do arg.build_core_arg().
This populates the _lib_handle appropriately.  C interface plan
routine now uses this to populate arg array for passing to
op_plan_core.