Support tuples of maps and 2-D dimension argument for Mat objects
Matrices and sparsities are now the union of the outer product of
pairs of maps, rather than just the outer product of a single pair of
maps.  The dimension argument specifies the shape of the local dense
block.  An integer argument means the dense block is square, thus
dim=2 gives a 2x2 dense block.  dim = (2,3) gives a 2x3 dense block.