Add Laplace demo.
This demo requires strong boundary conditions to be set. In order to implement
this, an interface to zero the rows of a matrix is added. A unit test for the
zeroing of rows is also added.

This Laplace demo is an intermediate step to one which uses Weak boundary
conditions, since Weak boundary conditions alone are not sufficient to make
the problem well-posed.