Working mass2d_vector demo
The hand-hacked code is removed from mass_vector_ffc and replaced
with calls to FFC which now supports assembly of vector fields in
PyOP2 code.

Instead of declaring the entire local matrix storage in the
generated code, the space for a single "element" is generated.
A loop over the dimensions being assembled over that contains the
addtos is generated unrolled.