Allow python GC to clean up core op_mat objects
Add a __del__ method to runtime_base and op_lib_core Mat objects.  The
former explicitly calls __del__ on the latter which just destroys the
C level matrix object and sets the pointer to NULL, the _lib_handle is
then set to None.

This fixes memory leaks in the case where we do:

    for i in range(...):
        mat = op2.Mat(...)