ffc_interface.compile form returns an ordered tuple of kernels
When compiling a form, check which domains (and hence integrals) are
present in the form and generate a kernel for each. The function
compile_form returns an ordered tuple of the kernels for
(cell, interior_facet, exterior_facet) where each of those is None if
the form does not contain a corresponding integral.

The rationale is that the user never actually needs the generated code
since it's only used to initialise a kernel anyway. Doing it this way we
don't need to expose the code and can wrap the logic for figuring out
which integrals are contained in the given form and which kernels we
therefore want to initialise inside compile_form.

Note: this doesn't yet work for multiple subdomains! There is also
further optimisation potential: Currently every kernel is initialised
with the generated code for all the integrals contained in the form
(since that is what FFC spits out). We could separate out the integrals
and only use those to initialise the kernels.

Update demos for kernels returned by compile_form: Combining the rhs
forms in weak_bcs_ffc demo requires passing both coefficients to either
par_loop for rhs assembly. This is due the way FFC generates code.