Add data_ro property to Dat objects
This property returns a readonly view of the data stored in a Dat (the
numpy array has the WRITEABLE flags set to false).  The programmer
promises that she only wants to look at the data and will not modify
values: this allows us to avoid transfers back to the device.