Add more tests of Global objects
This adds more tests for Globals in direct loops for 1D and 2D
globals.  We exercise reductions of all kinds (INC, MIN and MAX) for
different initial conditions.  In particular, this exercises the
correctness of reduction code for the different cases where either the
initial value of the Global, or a value in the Dat, will be the
correct final answer.

We also exercise resetting the data in a global in between par_loop
invocations.