Set all values of tracer in initial conditions par_loop
Because the tracer is accessed with WRITE, we must set all the values
in the par_loop.  For the sequential backend, this made no difference
because the data was zeroed when initialising the Dat.  For backends
that target a device, we only ensure that data is allocated on the
device when we encounter a WRITE only Dat in a par_loop.  We don't
upload data from the host, and so the initial values of the tracer Dat
were wrong.