Fix Plan cache key for matrices
Previously, the key for a matrix argument contained:

    as_tuple(arg.idx)

however, for matrices this is actually an IterationIndex object.  A
new IterationIndex object is instantiated each time we encounter a
parloop, so the key would be different.

Instead, unroll the index into the type signature (an IterationIndex)
and the indices the IterationIndex objects refer to.  We now expect
the multiple Plan with same matrix arguments test to pass, so remove
the previous expected failure.