Move fromhdf5 method into BackendSelector metaclass
Rather than having a separate metaclass for an object that needs a
fromhdf5 method, just give that method to all objects and raise an
exception if it doesn't exist and is invoked by the user.  This is in
preparation for the next patch which will allow us to instantiate the
correct backend-specific object in backend base classes.