First go at Argument conversion.
This sometimes violates ListTensor Assumption 1, which is that the
shape of every operand of a ListTensor is the same.

The shape is only perceived to be different because the
MixedElement was restricted by a FixedIndex under an Indexed,
where the Indexed fooled it into thinking that the Argument had
the shape specified by the Indexed's indices.