Add flag plot_convergence plotting convergence history to file
For each solve, the convergence history is plotted to a file named
reshist_N.png where N is a global counter counting solves in the Solver
class. A prefix can be specified via the plot_prefix parameter
(defaults to the empty string).

plot_convergence has no effect if monitor_convergence is not set.