Use our own MPI_COMM everywhere
We default to MPI_COMM_WORLD, if we see a halo we check to see if this
is the first one we've seen.  If so, use the communicator we got there
and subsequently check that all halos are defined on the same
communicator.

This may be wrong if halos actually live on a different communicator
than the one we want.  PETSc objects probably still don't live in the
right place.