Add uid initialiser to Dat objects
We tag the halo exchange of a Dat with its id slot.  Previously we
would just use the Dat._globalcount to get a unique value.  However,
this assumes that every process declares the same number of Dats in
the same order.  This does not always happen.  To fix this, allow the
user to specify the id explicitly.