Added logging of floating point operations
To log the number of floating point operations in the par_loop, take the kernel code and replace all floating point variables (float and double) by the type LoggedDouble in the file loggedtypes.hh. This type mimics the functionality of double, but increases a counter variable (implemented as a static class variable) whenever a floating point operation (+,-,*,/) is carried out. Then compile the kernel and run it on one element of the iteration set and return the logged number of FLOPs.
This currently only works if only basic floating point operations are used in the kernel (+,-,*,/). For example calling the 'pow()' or 'exp()' function in the kernel will fails, since they are not defined for the overloaded type. Furthermore, allowed kernel argument types are currently restricted to be Dat's or Global's.
The name of the g++ compiler has been hardcoded, and this needs to be fixed.