Allocated contiguous block of memory
In some PyOP2 ParLoops (more specifically in the BandedMatrix class in the firedrake-helmholtzsolver repository) the data in a cell might be accessed in a 'flatened' way, i.e. the user might access x[0][n*i+j] instead of x[i][j] in the kernel. This works fine in PyOP2, but in the C++ kernel that counts the floating point operations, the data array of type LoggedDouble was not allocated as one chunk of memory. This commit fixes this problem.