WIP: Adapting codegen to match shapes with TSFC
This is attempting to remove the need to call the (soon-to-be-removed,
see inducer/loopy#327) loopy functionality that allows us to match
caller and callee kernel shapes.

TODO: Would it be better to deliver loopy flattened arguments?
Probably not.

TODO: We presently don't make the distinction between a vector Dat
with a single component (shape (1,)) and a scalar Dat (shape ()).
Instead both advertise that they have shape (1,). This is problematic
because on the TSFC side we do make that distinction. Probably this
means that scalar Dats should have shape ().