Only delete compcomm on icomm if it exists
Within the MPI implementation, an inner communicator can optionally
contain a compilation commucator. An indentation error meant that the
compilation attribute was unconditionally deleted, causing incorrect
MPI behaviour if the attribute had never been set in the first
place. On MPICH, this doesn't seem to raise any errors, but with
OpenMPI it causes errors like:

    mpi4py.MPI.Exception: MPI_ERR_OTHER: known error not in list